requirements.txt
python-dotenv>=1.0.0
openai>=1.3.0
pyzotero
colorama>=0.4.6
rich>=13.0.0
requests>=2.31.0
qdrant-client[fastembed]>=1.14.2
